- Reworked NetworkService from Name-prefixed string-based packets to Data Packets - Changed type of all ids to long(signed int64)

+++ Client/ViewModels/ChatViewModel.cs
@@ -16,7 +16,7 @@ namespace Client.ViewModels
             _chat.PropertyChanged += ChatOnPropertyChanged;
         }
 
-        public int Id => _chat.Id;
+        public long Id => _chat.Id;
         public string Title
         {
             get => _chat.User.Fullname;
